Java Script working for the lbox
the lbox now has working replicants, the way it's set is that one script writes to the replicant while the other reads from it and writes to the on screen element. The plan will be to add these for every element on screen. As a warning these elements are not sanitized so you can casually paste raw html in them.